+++ CHANGES.md
@@ -9,3 +9,4 @@
 * Moved App/System/CurrentPage
 * Moved App/Services to App/ServiceModules
 * class Service to ServiceModule
+* Removed simple service class